feat: 게임 페이지 분리
- 모드마다 컴포넌트로 분리
