+++ packages/frontend/src/components/Gartic/Gartic.component.tsx
@@ -1,0 +1,5 @@
+const Gartic = () => {
+  return <></>;
+};
+
+export default Gartic;
